requirements.txt
fastapi
uvicorn
langchain
faiss-cpu
sentence-transformers
wikipedia
python-docx
PyPDF2
pdfplumber
python-multipart